Introduced GUIAnyParam
The purpose of GUIAnyParam is to encapsulate a vst param or a jmb param or a single value (if no param available). This allows writing generic views that do not care whether the value is coming from vst, jmb or simple value.

The intent is to properly implement what the __internal__registerRawVstControl  method wanted to achieve with added bonus that it extends it to jmb parameters as well!